Plugin/1-Wire ================== * Creation (only OWFS mode supported, and not tested)
shared/yPluginApi
==================
* Add Illumination (lux) historizer
* Add Illumination (lux) capacity
* Add lux unit
* Counter historizer : change value type from int to long long

+++ trunk/sources/shared/shared/plugin/yPluginApi/historization/Counter.cpp
@@ -6,7 +6,7 @@ namespace shared { namespace plugin { namespace yPluginApi { namespace historiza
 {
 
    CCounter::CCounter(const std::string& keywordName, const EKeywordAccessMode& accessMode, const EMeasureType& measureType, typeInfo::CIntTypeInfo & additionalInfo)
-      :CSingleHistorizableData<int>(keywordName, CStandardCapacities::Counter, accessMode, 0, measureType, additionalInfo)
+      :CSingleHistorizableData<long long>(keywordName, CStandardCapacities::Counter, accessMode, 0, measureType, additionalInfo)
    {
    }
